+++ core/src/main/java/com/yahoo/ycsb/measurements/exporter/TextMeasurementsExporter.java
@@ -27,6 +27,8 @@ import java.io.OutputStreamWriter;
 public class TextMeasurementsExporter implements MeasurementsExporter {
   private final BufferedWriter bw;
 
+  //TODO
+
   public TextMeasurementsExporter(OutputStream os) {
     this.bw = new BufferedWriter(new OutputStreamWriter(os));
   }